+++ NEWS.md
@@ -12,6 +12,11 @@ plumber 0.5.0
 
 ### Breaking changes
 
+* Filters that want to forward execution onto the next filter or endpoint must
+  return the result of `forward()`. Simply calling the `forward()` function at some
+  point during the filter will not work anymore. This helps enable the handling of
+  promises from within filters, hooks, and endpoints.
+
 * Plumber's swagger definition is now defined using
   [OpenAPI 3](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md),
   upgrading from Swagger Specification. ([#365](https://github.com/trestletech/plumber/pull/365))